Examined abstract and changed something.
however, I think that by "extended abstract" something more is needed:
Something with a proper paper structure: authors, abstract, intro,
maybe conclusions and a few references in 1-2 pages, not only an
abstract. At least it should be in proper paper format, maybe the
definitive one. This abstract will be extended to the whole paper
later on.